+++ src/Exception/FileAlreadyExistsException.php
@@ -2,8 +2,6 @@
 
 namespace Pontedilana\PhpWeasyPrint\Exception;
 
-use InvalidArgumentException;
-
-class FileAlreadyExistsException extends InvalidArgumentException
+class FileAlreadyExistsException extends \InvalidArgumentException
 {
 }