+++ tests/Unit/ImageTest.php
@@ -1,12 +1,14 @@
 <?php
 
-namespace Tests\Unit\Pontedilana\PhpWeasyPrint;
+namespace Pontedilana\PhpWeasyPrint\Tests\Unit;
 
 use PHPUnit\Framework\TestCase;
 use Pontedilana\PhpWeasyPrint\Image;
 
 /**
  * @covers \Pontedilana\PhpWeasyPrint\Image
+ *
+ * @deprecated 1.0.0 Image class is deprecated and will be removed
  */
 class ImageTest extends TestCase
 {